disable video recording on snapshots and delete videos on successful tests

+++ e2e/support/config.js
@@ -129,15 +129,10 @@ const defaultConfig = {
     // https://docs.cypress.io/guides/guides/screenshots-and-videos#Delete-videos-for-specs-without-failing-or-retried-tests
     on("after:spec", (spec, results) => {
       if (results && results.video) {
-        // Do we have failures for any retry attempts?
-        const failures = results.tests.some(test =>
-          test.attempts.some(attempt => attempt.state === "failed"),
-        );
-        if (!failures) {
-          // delete the video if the spec passed and no tests retried
-          if (fs.existsSync(results.video)) {
-            fs.unlinkSync(results.video);
-          }
+        // Do we have test failures?
+        if (results && results.video && results.stats.failures === 0) {
+          // delete the video if the spec passed
+          fs.unlinkSync(results.video);
         }
       }
     });
